Split sort back into Ascending / Descending variants.  Use header row is still a popup choice.
Added Select column command to block select the current column.  (Works in justified and collapsed modes)  When used with justified columns, this allows a variety of editing operations, such as quickly adding, moving, and deleting columns.

The selection regions are now saved & restored around edit operations.

Fixed an exception when doing an operation when the last line of the file was selected, and the line was empty.