Improve parsing of escaped double quotes
Per the RFC, "" is treated as an escaped quote.
Unfortunately this leads to quoted empty cells, e.g. "","","" being treated as escaped quotes instead of empty cells.

A more careful reading of the spec indicates that this is only valid when already inside double quotes, and the parser has been updated to reflect this.

Also updates the readme to clarify that quoted newlines in cells are not supported.